test: add portfolio e2e tests
